+++ CHANGELOG.md
@@ -1,5 +1,24 @@
 # Changelog
 
+## [1.1.0] - 2022-01-30
+
+### Added
+
+- Adds a local cache of the species taxonomy to avoid repeated queries to the NCBI.
+
+
+## [1.0.3] - 2022-01-27
+
+### Changed
+
+- Show versions of the main PTC script and the PTC JAR file.
+
+## [1.0.2] - 2022-01-27
+
+### Fixed
+
+- Fix `docker run` command when host path contains spaces.
+
 ## [1.0.1] - 2022-01-24
 
 ### Changed